+++ user-service/src/main/kotlin/com/toy/userservice/account/adapter/infrastructure/auth/JwtAuthenticationFilter.kt
@@ -1,6 +1,5 @@
-package com.toy.userservice.account.adapter.infrastructure.security
+package com.toy.userservice.account.adapter.infrastructure.auth
 
-import com.toy.userservice.account.adapter.infrastructure.security.CustomUserDetailsService
 import io.github.oshai.kotlinlogging.KotlinLogging
 import jakarta.servlet.FilterChain
 import jakarta.servlet.http.HttpServletRequest
