+++ user-service/src/main/kotlin/com/toy/userservice/account/adapter/infrastructure/auth/JwtProperties.kt
@@ -1,4 +1,4 @@
-package com.toy.userservice.account.adapter.infrastructure.security
+package com.toy.userservice.account.adapter.infrastructure.auth
 
 import org.springframework.boot.context.properties.ConfigurationProperties
 import java.time.Duration
