+++ user-service/src/main/kotlin/com/toy/userservice/account/adapter/infrastructure/auth/JwtProvider.kt
@@ -1,4 +1,4 @@
-package com.toy.userservice.account.adapter.infrastructure.security
+package com.toy.userservice.account.adapter.infrastructure.auth
 
 import com.toy.userservice.account.domain.model.User
 import io.github.oshai.kotlinlogging.KotlinLogging
