+++ user-service/src/main/kotlin/com/toy/userservice/account/adapter/infrastructure/auth/SecurityConfig.kt
@@ -1,6 +1,5 @@
-package com.toy.userservice.account.adapter.infrastructure.security
+package com.toy.userservice.account.adapter.infrastructure.auth
 
-import com.toy.userservice.account.adapter.infrastructure.security.JwtProperties
 import org.springframework.boot.context.properties.EnableConfigurationProperties
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
