+++ user-service/src/test/kotlin/com/toy/userservice/account/adapter/infrastructure/auth/JwtPropertiesTest.kt
@@ -1,6 +1,5 @@
-package com.toy.userservice.account.adapter.`in`.web.common.security
+package com.toy.userservice.account.adapter.infrastructure.auth
 
-import com.toy.userservice.account.adapter.infrastructure.security.JwtProperties
 import org.assertj.core.api.Assertions.*
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Nested
